Updated name to ProkFunFind
