--- GutFunFind/cluster/DBSCAN/__init__.py
@@ -1,71 +1,0 @@
-from typing import List
-import numpy as np
-from sklearn.cluster import DBSCAN
-
-from GutFunFind.read import Genome
-# from GutFunFind.toolkit.utility import *
-
-
-def DBSCANCluster(
-        pos: List[int],
-        weight: List[float] = None,
-        eps: float = 2,
-        min_samples: float = 0.9) -> List[int]:
-    """Wrapper function to run DBSCAN clustering
-
-       Arguments:
-           pos:
-           weight:
-           eps: 
-           min_samples:
-
-       Returns:
-           list of cluster labels for genes 
-    """
-    pos = np.array(pos).reshape(-1, 1)
-    cluster = DBSCAN(
-        eps=eps, min_samples=min_samples).fit(pos, sample_weight=weight)
-    return list(cluster.labels_)
-
-
-def pipeline(config: dict, genome_object, detect_tools: set) -> Genome:
-    """Main pipeline function for DBSCAN clustering analysis
-
-       Arguments:
-           config: A config dictionary 
-           genome_object: A GutFunFind.annotate.Genome object
-           detect_tools: list of gene hit detection tools used
-       Returns: 
-           genome_object: GutFunFind.annotate.Genome object with cluster info
-    """
-    for ct in genome_object.contigs.values():
-
-        # obtain weight for genes with blast attributes and the postion in the
-        # contig(0-based)
-        poslist, weightlist, genelist = [], [], []
-        for detect_tool in detect_tools:
-            for i, gene in enumerate(ct.genes.values()):
-                if hasattr(gene, detect_tool):
-                    poslist.append(i)
-                    genelist.append(gene)
-                    weightlist.append(
-                        getattr(
-                            getattr(
-                                gene,
-                                detect_tool),
-                            "orthoID_weight"))
-
-        # if there is a hit in the contig, run DBSCANCluster
-        if len(poslist) >= 1:
-            cl = DBSCANCluster(pos=poslist,
-                               weight=weightlist,
-                               eps=float(config['DBSCAN']['cluster.eps']),
-                               min_samples=float(config['DBSCAN'][
-                                                 'cluster.min_samples']))
-
-            for gene, cl_label in zip(genelist, cl):
-                if cl_label == -1:
-                    cl_label = "NA"
-                setattr(gene, "DBSCAN", cl_label)
-
-    return genome_object
