+++ ProkFunFind/tests/test_examine.py
@@ -1,0 +1,6 @@
+from ProkFunFind.read import *
+from Bio.SeqFeature import FeatureLocation
+
+g1 = Genome(contigs={'contig1': Contig(id='contig1', seq='ATTAGAGAACGGTCGTAACATTATCGGTGGTTCTCTAACTACTATCAGTACCCATGACTCGACTCTGCCGCAGCTACCTATCGCCTGAAAGCCAGTTGGTGTTAAGGAGTGCTCTGTCCAGGACAACACGCGTAGTGAGAGTTACATGTTCGTTGGGTTCTCCCGACTCGGACCTGAGTCGACCAAGGACCCACTCGAGCTCTGAGCCCCACTGTCGAGAAGTATGTATCTCGCTCCCGCAGCTTGCCAGCACTTTCAGTATCATGGGGCCCATGGTTGAATGACTCCTATAACGGACTTCGACATGGCAAAATCCCCCCCTTGCAACTTCTAGAGGAGAAGAGTACTGACTTGAGCGCTCCCAGCACAACGGCCAAGGAAGTCTCCAATTTCTTGTTTCCGAATGACACGCGTCTCCTTGCGGGTAAATCGCCGACCGCAGAACTTAGGAGCCAGGGGGAACAGATAGGTCTAATTAGGTTAAGGGAGTAAGTCCTCGGATGGTTCAGTTGTAACCATATACTTACGCTGGAACTTCTCCGGCGAATTTTTACTGTCACCAACCACGAGATTTGAGGTAAACCAATTGAGCACATAGTCGCGCTATCCGACAATCTCCAAATTATAACATACCGTTCCATGAAGGCCAGAGTTACTTACCGGCCCTTTCCATGCGCGCGCCATACCCCCCCAGTTCCCCGGTTATCTCTCCGAGGAGAGAGTGAGCGATCCTCCGTTAACATATTCTTACGTATGACGTAGCTATGTATTTTGCAGAGGTAGCGAACGGGTTTGACACTTCACAGATAGTGGGGATCCGGGCAAAGGGCGTATAATTGCGGTCCAACATAGGCGTAAACTACGATGGCACCAACTCAGTCGCAGCTCGTGCGCCGTGAATAACGTACTCATCCCAACTGATTCTCGGCAATCTACGGAGCGACATGATTATCAACAGCTGTCTAGCAGTTCTAATCTTTTGCCATCGTCGTAAAAGCCTCCAAGAGATTGATCATACCTATCGGCACAGAAGTGACACGACGCCGATGGGTAGCGGACTTTTGGTCAACCACAATTCGCTAGGGGACAGGTCCTGCGGCGTACATCACTTTGTATGTGCAACCAGCCCAAGTGGGGCCAGGCAAGACTCAGCTGGTTCCTGTGTTAGCTCGAGGCTAGGGATGACAGCTCTTTAAACATAGGCTGGGGGCGTCGAACGGTCGAGAAGCTCATAGTACCTCGGGTACCAACTTACTCAGGCTATTGCTTGAAGCTGTACTATTTCAGGGGGGGAGCGCTGATGGTCTCTTCTTCTGATGACTCAACTCGCAAGGGTCGTGAAGTCGGTTCCTTCAATGGTTAAAAATCAAAGGCTCACTGTGCAGACTGGAGCGCCCATCTAGCGGCTCGCGTCTCGAATGCTCGGTCCCCTTTGTCATTCCGGATAAATTCATACCCCTCATTCACTAGCTTGCGAAGTCTACATTGGTATATGAATCCGACCTAGAAGAGGGCACTTAAAATTGGGAGTGGTTGATGCTCTATACTCCATTCGGTTTTTTCGTGCATCACCGCGAGAGGCTGACAAGGGTTTGACATTGAGTAGCAAGGCACTTCCGGACTCAATGAAGGGCCGGGAAAGGTACGCGCGTGGTATGGGAGGATCAAGGAGCCAATACAAAGGCTTCATCCTCACTCGCATGGAGGCAAACGCAGAACAATGGTTACTATTTCGATACGTGAAACATGTCCCACGGTAGCCCAAAGACTTAAGAGTCTATCACCCCTAGGGCCCTTTCCCGGATATAAACGCCAGGTTGAATCCGCATTTGGAGCTACGATGGATGAGTCTGGGTGGAGCGCGCCCCATTTATACCGTGAGTAGGGTCGACCAAGAACCGCAAGATGCGTCGGTGTACAAATAATTGTCAACAGACCGTCGTGTTTTGAAAATGGTACCAGCATCTTCGGGCGGTCTCAATCAAGCATGGATTACGGTG', genes={'gene1':Gene(contig='contig1', id='gene1', type='CDS', qualifiers='', location=FeatureLocation(1,150))})})
+
+print(g1.contigs)
