--- bin/run_GutFunFind.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-
-from GutFunFind.__main__ import main
-
-if __name__ == '__main__':
-    main()
