+++ docs/source/help.txt
@@ -1,4 +1,4 @@
-usage: run_GutFunFind.py [-h] {rep,pan} ...
+usage: prokfunfind.py [-h] {rep,pan} ...
 
 Identify genes related function of interest
