+++ docs/source/index.rst
@@ -1,11 +1,11 @@
-.. GutFunFind documentation master file, created by
+.. ProkFunFind documentation master file, created by
    sphinx-quickstart on Fri Jul 24 19:50:07 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to GutFunFind's documentation!
+Welcome to ProkFunFind's documentation!
 ======================================
-  GutFunFind is a pipeline that can be used to predict genes related to function of interest.
+  ProkFunFind is a pipeline that can be used to predict genes related to function of interest.
   Criteria for function detection include presence of detected genes, and genomic co-localization.
 
   **In order to identify the correct genes, the user needs:**
@@ -18,7 +18,7 @@ Welcome to GutFunFind's documentation!
      - ...
 
    - Choose the correct bait sequences/domain signatures/HMM protein profiles for the components of interest
-   - Specify the definition of gene cluster and cluster method 
+   - Specify the definition of gene cluster and cluster method
 
      - DBSCAN
      - ...
@@ -36,3 +36,4 @@ Welcome to GutFunFind's documentation!
    outputs
    run_parallel
    todo
+   tutorial/
