+++ docs/source/tutorial/index.rst
@@ -1,0 +1,25 @@
+ProkFunFind Tutorial
+====================
+
+The purpose of this tutorial is to walk you through the process of setting up
+and executing searches using the ProkFunFind tool. The tutorial will use the
+example of searching for a gene cluster that encodes the metabolic pathway for
+Equol production (for the actual study where we analyzed this gene cluster
+see PMID: 35247986).
+
+**Tutorial Requirements**
+
+The tutorial materials have been packaged into their own git repository: URL # ADD GIT REPO LINK
+The tutorial materials include all of the input and output files that are needed
+or would be generated while running the tutorial steps. It has also been designed
+to be modular, so each tutorial section can be run on its own without going
+through any of the other sections first.
+
+At minimum you will need to install ProkFunFind and download the tutorial
+materials to run a majority of the sections. The ProkFunAnnotate tutorial
+covers how to use our annotation pipeline and requires multiple additional
+programs to be installed and configured. All of the information related to
+installation of ProkFunAnnotate and ProkFunFind and the setup of the tutorial
+materials can be found in the "Installation and Materials" tutorial section.
+
+**Tutorial Materials**
